+++ kmean/src/mapreduce/CanopyCombiner.java
@@ -2,7 +2,7 @@
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import utils.VectorDoubleWritable;
@@ -10,9 +10,9 @@ import canopy.Canopy;
 
 public class CanopyCombiner
 		extends
-		Reducer<IntWritable, Iterable<VectorDoubleWritable>, IntWritable, Canopy> {
+		Reducer<LongWritable, Iterable<VectorDoubleWritable>, LongWritable, Canopy> {
 
-	public void reduce(IntWritable key, Iterable<VectorDoubleWritable> values,
+	public void reduce(LongWritable key, Iterable<VectorDoubleWritable> values,
 			Context context) throws IOException {
 		Canopy canopy = new Canopy();
 		for (VectorDoubleWritable point : values) {
