+++ kmean/src/mapreduce/CanopyMapper.java
@@ -3,7 +3,7 @@ package mapreduce;
 import java.io.IOException;
 import java.util.ArrayList;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
 
@@ -13,7 +13,7 @@ import clusterer.CanopyClusterer;
 import utils.VectorDoubleWritable;
 
 public class CanopyMapper extends
-		Mapper<IntWritable, Text, IntWritable, VectorDoubleWritable> {
+		Mapper<LongWritable, Text, LongWritable, VectorDoubleWritable> {
 	private VectorDoubleWritable point = null;
 	protected CanopyClusterer canopyClusterer = new CanopyClusterer();
 	private ArrayList<CanopyCluster> canopies = new ArrayList<CanopyCluster>();
@@ -21,7 +21,7 @@ public class CanopyMapper extends
 
 
 	@Override
-	protected void map(IntWritable key, Text values, Context context)
+	protected void map(LongWritable key, Text values, Context context)
 			throws IOException, InterruptedException {
 		System.out.println(values.toString());
 		point = new VectorDoubleWritable(values);
@@ -40,7 +40,7 @@ public class CanopyMapper extends
 	protected void cleanup(Context context) throws IOException,
 			InterruptedException {
 		for (CanopyCluster canopy : canopies) { 
-			context.write(new IntWritable(canopy.getId()),
+			context.write(new LongWritable(canopy.getId()),
 					canopy.getCentroid());
 		}
 		super.cleanup(context);
